requirements.txt
streamlit
pandas
feedparser
python-dotenv
requests